Fixes #97 - Changes the structure of the popup
