+++ app/popup/src/components/App/reset.css
