+++ app/popup/src/components/Header/AccountView/index.js
