--- app/popup/src/components/Main/Main.css
@@ -1,6 +1,0 @@
-.mainContainer {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    width: 100%;
-}
