--- app/popup/src/index.css
@@ -1,23 +1,0 @@
-@import "./reset.css";
-
-@import url('https://fonts.googleapis.com/css?family=Roboto:300,400,500,700');
-@import url('https://fonts.googleapis.com/css?family=Source+Code+Pro:300,400,500,700');
-
-html {
-    height: 595px;
-    width: 420px;
-}
-
-body, #root {
-    height: 100%;
-    width: 100%;
-    overflow: hidden;
-    font-family: 'Roboto', sans-serif;
-    background: linear-gradient(180deg, #fff, #efefef);
-    cursor: default;
-    user-select: none;
-}
-
-:focus {
-    outline: none;
-}
