+++ app/popup/src/pages/Queue/CreateSmartcontract/index.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import './CreateSmartContract.css';
 
-import { popup} from '../../index.js';
-import { updateConfirmations } from '../../reducers/confirmations';
+import { popup } from 'index.js';
+import { updateConfirmations } from 'reducers/confirmations';
 
 class CreateSmartContract extends Component {
     constructor(props){
