+++ app/popup/src/pages/Queue/index.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { connect } from 'react-redux';
 import './Queue.css';
 
-import { CONFIRMATION_TYPE } from '../../extension/constants';
+import { CONFIRMATION_TYPE } from 'extension/constants';
 
-import Send from './SendTRX.js';
-import CreateSmartContract from './CreateSmartContract.js';
+import Send from './SendTRX';
+import CreateSmartContract from './CreateSmartContract';
 
 class Queue extends React.Component {
     renderConfirmationType() {
@@ -14,8 +14,10 @@ class Queue extends React.Component {
         switch(confirmation.type) {
             case CONFIRMATION_TYPE.SEND_TRON:
                 return <Send />;
+
             case CONFIRMATION_TYPE.CREATE_SMARTCONTRACT:
                 return <CreateSmartContract/>;
+                
             default:
                 return (
                     <span>
